+++ frontend/src/components/placeContentBox/placeContentBox.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+const PlaceContentBox = ({ places }) => {
+  // places : place id, place images
+  return <div>hi</div>;
+};
+
+export default PlaceContentBox;
